Started to add Javadoc comments.
Made the allocation of the input in the class Player caseinsensitiv.

Changed visibility of Map.shuffleSpawnPoints() to public so the spawnpoints can be randomized for each game now.

Changed in BasicSnake the constructor. At the start of a game a space be will be set as input for all players so the last input from a previous game will be overwritten and everyone has the same start direction.

In Lobby the final state of the attribute gamemode was removed to fix several time problems. Like problems with the countdown and overall roundtime. Also added map.shuffleSpawnPoints() to randomize spawnpoints. Now everytime you press start in a Lobby you get a new BasicSnake instance with shuffled spawnpoints.